+++ src/datos.py
@@ -1,8 +1,7 @@
 import json
 import os
-import datos
-from utils import limpiar_pantalla
-from reportes import exportar_clientes_csv, exportar_reservas_csv, exportar_hoteles_csv
+
+
 
 
 
@@ -104,35 +103,3 @@ if __name__ == "__main__":
     print("RUTA_ARCHIVO_RESERVAS:", RUTA_ARCHIVO_RESERVAS)
 
     
-def exportar_datos_csv(hoteles: list, clientes: list, reservas: list) -> None:
-   
-    """
-    Función principal para gestionar la exportación de todos los datos a CSV.
-    Maneja la lógica de rutas y llama a las funciones de exportación.
-    Pre: Recibe las listas de hoteles, clientes y reservas.
-    
-    Post: Exporta los datos a archivos CSV en las rutas especificadas.
-    """
-    limpiar_pantalla()
-    print("=" * 40)
-    print("   --- Exportando datos a CSV ---     ".center(40, " "))
-    print("=" * 40)
-    
-    try:
-        # 1. Definir las rutas (usando datos.RUTA_DATA)
-        ruta_clientes_csv = os.path.join(datos.RUTA_DATA, "clientes_export.csv")
-        ruta_reservas_csv = os.path.join(datos.RUTA_DATA, "reservas_export.csv")
-        ruta_hoteles_csv = os.path.join(datos.RUTA_DATA, "hoteles_export.csv")
-        ruta_habitaciones_csv = os.path.join(datos.RUTA_DATA, "habitaciones_export.csv")
-
-        # 2. Llamar a las funciones de exportación (que están en este mismo archivo)
-        exportar_clientes_csv(clientes, ruta_clientes_csv)
-        exportar_reservas_csv(reservas, ruta_reservas_csv)
-        exportar_hoteles_csv(hoteles, ruta_hoteles_csv, ruta_habitaciones_csv)
-        
-        print(f"\n¡Datos exportados exitosamente en la carpeta '{datos.RUTA_DATA}'!")
-    
-    except Exception as e:
-        print(f"\nOcurrió un error general al exportar: {e}")
-    
-    input("\nPresione Enter para continuar...")
